move read code into read.rs
