+++ src/lib.rs
@@ -93,12 +93,15 @@ mod map;
 mod matchers;
 mod matchtype;
 
+#[cfg(feature = "std")]
+mod read;
+
 #[cfg(feature = "alloc")]
 use alloc::vec::Vec;
 #[cfg(feature = "std")]
 use std::fs::File;
 #[cfg(feature = "std")]
-use std::io::{self, Read, Seek};
+use std::io::{self, Read};
 #[cfg(feature = "std")]
 use std::path::Path;
 
@@ -109,9 +112,10 @@ use map::{WrapMatcher, WrapReadMatcher};
 
 use map::MATCHER_MAP;
 
+#[cfg(feature = "std")]
+pub use crate::read::*;
 /// All the supported matchers categorized and exposed as functions
 pub use matchers::*;
-
 pub use matchtype::*;
 
 /// Infer allows to use a custom set of `Matcher`s for infering a MIME type.
@@ -160,44 +164,6 @@ impl Infer {
         self.iter_matchers().find(|kind| kind.matches(buf)).copied()
     }
 
-    /// Returns the file type of the data in the reader.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use std::fs;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> std::io::Result<()> {
-    ///     let info = infer::Infer::new();
-    ///     let mut f = File::open("testdata/sample.jpg")?;
-    ///     let kind = info.get_read(&mut f).unwrap().expect("file type is known");
-    ///     assert_eq!(kind.mime_type(), "image/jpeg");
-    ///     assert_eq!(kind.extension(), "jpg");
-    ///     Ok(())
-    /// }
-    /// ```
-    #[cfg(feature = "std")]
-    pub fn get_read<R>(&self, r: &mut R) -> io::Result<Option<Type>>
-    where
-        R: Read + Seek,
-    {
-        let mut res_value: Option<Type> = None;
-
-        for kind in self.iter_matchers() {
-            let match_res = kind.matches_read(r)?;
-            if match_res {
-                res_value = Some(*kind);
-                break;
-            }
-
-            r.rewind().ok();
-        }
-
-        Ok(res_value)
-    }
-
     /// Returns the file type of the file given a path.
     ///
     /// # Examples
@@ -227,33 +193,6 @@ impl Infer {
             .any(|kind| kind.extension() == extension && kind.matches(buf))
     }
 
-    /// Determines whether data from read is of given extension.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_read`](./fn.is_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_read<R>(&self, r: &mut R, extension: &str) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        let mut res_value: bool = false;
-
-        for kind in self.iter_matchers() {
-            if kind.extension() == extension {
-                let match_res = kind.matches_read(r)?;
-                if match_res {
-                    res_value = true;
-                    break;
-                }
-
-                r.rewind().ok();
-            }
-        }
-
-        Ok(res_value)
-    }
-
     /// Determines whether a buffer is of given mime type.
     ///
     /// # Examples
@@ -262,32 +201,6 @@ impl Infer {
     pub fn is_mime(&self, buf: &[u8], mime_type: &str) -> bool {
         self.iter_matchers()
             .any(|kind| kind.mime_type() == mime_type && kind.matches(buf))
-    }
-
-    /// Determines whether data from reader is of given mime type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_mime_read`](./fn.is_mime_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_mime_read<R>(&self, r: &mut R, mime_type: &str) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        let mut res_value: bool = false;
-        for kind in self.iter_matchers() {
-            if kind.mime_type() == mime_type {
-                let match_res = kind.matches_read(r)?;
-                if match_res {
-                    res_value = true;
-                    break;
-                }
-
-                r.rewind().ok();
-            }
-        }
-
-        Ok(res_value)
     }
 
     /// Returns whether an extension is supported.
@@ -341,19 +254,6 @@ impl Infer {
         self.is_type(buf, MatcherType::App)
     }
 
-    /// Determines whether data is an application type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_app_read`](./fn.is_app_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_app_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::App)
-    }
-
     /// Determines whether a buffer is an archive type.
     ///
     /// # Examples
@@ -361,19 +261,6 @@ impl Infer {
     /// See [`is_archive`](./fn.is_archive.html).
     pub fn is_archive(&self, buf: &[u8]) -> bool {
         self.is_type(buf, MatcherType::Archive)
-    }
-
-    /// Determines whether data from reader is an archive type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_archive_read`](./fn.is_archive_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_archive_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::Archive)
     }
 
     /// Determines whether a buffer is an audio type.
@@ -385,19 +272,6 @@ impl Infer {
         self.is_type(buf, MatcherType::Audio)
     }
 
-    /// Determines whether data from reader is an audio type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_audio_read`](./fn.is_audio_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_audio_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::Audio)
-    }
-
     /// Determines whether a buffer is a book type.
     ///
     /// # Examples
@@ -405,19 +279,6 @@ impl Infer {
     /// See [`is_book`](./fn.is_book.html).
     pub fn is_book(&self, buf: &[u8]) -> bool {
         self.is_type(buf, MatcherType::Book)
-    }
-
-    /// Determines whether data from reader is a book type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_book_read`](./fn.is_book_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_book_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::Book)
     }
 
     /// Determines whether a buffer is a document type.
@@ -429,19 +290,6 @@ impl Infer {
         self.is_type(buf, MatcherType::Doc)
     }
 
-    /// Determines whether data from reader is a document type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_document_read`](./fn.is_document_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_document_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::Doc)
-    }
-
     /// Determines whether a buffer is a font type.
     ///
     /// # Examples
@@ -449,19 +297,6 @@ impl Infer {
     /// See [`is_font`](./fn.is_font.html).
     pub fn is_font(&self, buf: &[u8]) -> bool {
         self.is_type(buf, MatcherType::Font)
-    }
-
-    /// Determines whether data from reader is a font type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_font_read`](./fn.is_font_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_font_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::Font)
     }
 
     /// Determines whether a buffer is an image type.
@@ -473,19 +308,6 @@ impl Infer {
         self.is_type(buf, MatcherType::Image)
     }
 
-    /// Determines whether data from reader is an image type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_image_read`](./fn.is_image_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_image_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::Image)
-    }
-
     /// Determines whether a buffer is a video type.
     ///
     /// # Examples
@@ -493,19 +315,6 @@ impl Infer {
     /// See [`is_video`](./fn.is_video.html).
     pub fn is_video(&self, buf: &[u8]) -> bool {
         self.is_type(buf, MatcherType::Video)
-    }
-
-    /// Determines whether data from reader is a video type.
-    ///
-    /// # Examples
-    ///
-    /// See [`is_video_read`](./fn.is_video_read.html).
-    #[cfg(feature = "std")]
-    pub fn is_video_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::Video)
     }
 
     /// Determines whether a buffer is one of the custom types added.
@@ -528,15 +337,6 @@ impl Infer {
     /// ```
     pub fn is_custom(&self, buf: &[u8]) -> bool {
         self.is_type(buf, MatcherType::Custom)
-    }
-
-    /// Determines whether data from reader is one of the custom types added.
-    #[cfg(feature = "std")]
-    pub fn is_custom_read<R>(&self, r: &mut R) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        self.is_type_read(r, MatcherType::Custom)
     }
 
     /// Adds a custom matcher.
@@ -585,28 +385,6 @@ impl Infer {
         self.iter_matchers()
             .any(|kind| kind.matcher_type() == matcher_type && kind.matches(buf))
     }
-
-    #[cfg(feature = "std")]
-    fn is_type_read<R>(&self, r: &mut R, matcher_type: MatcherType) -> io::Result<bool>
-    where
-        R: Read + Seek,
-    {
-        let mut res_value: bool = false;
-
-        for kind in self.iter_matchers() {
-            if kind.matcher_type() == matcher_type && kind.supports_read_match() {
-                let match_res = kind.matches_read(r)?;
-                if match_res {
-                    res_value = true;
-                    break;
-                }
-
-                r.rewind().ok();
-            }
-        }
-
-        Ok(res_value)
-    }
 }
 
 impl Default for Infer {
@@ -630,31 +408,6 @@ static INFER: Infer = Infer::new();
 /// ```
 pub fn get(buf: &[u8]) -> Option<Type> {
     INFER.get(buf)
-}
-
-/// Returns the file type of the data in the reader.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.jpg")?;
-///     let kind = infer::get_read(&mut f).unwrap().expect("file type is known");
-///     assert_eq!(kind.mime_type(), "image/jpeg");
-///     assert_eq!(kind.extension(), "jpg");
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn get_read<R>(r: &mut R) -> io::Result<Option<Type>>
-where
-    R: Read + Seek,
-{
-    INFER.get_read(r)
 }
 
 /// Returns the file type of the file given a path.
@@ -690,30 +443,6 @@ pub fn is(buf: &[u8], extension: &str) -> bool {
     INFER.is(buf, extension)
 }
 
-/// Determines whether a buffer is of given extension.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.jpg")?;
-///     let jpg = infer::is_read(&mut f, "jpg").unwrap();
-///     assert!(jpg);
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_read<R>(r: &mut R, extension: &str) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_read(r, extension)
-}
-
 /// Determines whether a buffer is of given mime type.
 ///
 /// # Examples
@@ -724,29 +453,6 @@ where
 /// ```
 pub fn is_mime(buf: &[u8], mime_type: &str) -> bool {
     INFER.is_mime(buf, mime_type)
-}
-
-/// Determines whether data from reader is of given mime type.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.jpg")?;
-///     assert!(infer::is_mime_read(&mut f, "image/jpeg").unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_mime_read<R>(r: &mut R, mime_type: &str) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_mime_read(r, mime_type)
 }
 
 /// Returns whether an extension is supported.
@@ -783,29 +489,6 @@ pub fn is_app(buf: &[u8]) -> bool {
     INFER.is_app(buf)
 }
 
-/// Determines whether data from reader is an application type.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.wasm")?;
-///     assert!(infer::is_app_read(&mut f).unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_app_read<R>(r: &mut R) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_app_read(r)
-}
-
 /// Determines whether a buffer is an archive type.
 /// # Examples
 ///
@@ -815,28 +498,6 @@ where
 /// ```
 pub fn is_archive(buf: &[u8]) -> bool {
     INFER.is_archive(buf)
-}
-
-/// Determines whether data from reader is an archive type.
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.pdf")?;
-///     assert!(infer::is_archive_read(&mut f).unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_archive_read<R>(r: &mut R) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_archive_read(r)
 }
 
 /// Determines whether a buffer is an audio type.
@@ -852,29 +513,6 @@ pub fn is_audio(buf: &[u8]) -> bool {
     INFER.is_audio(buf)
 }
 
-/// Determines whether data from reader is an audio type.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.mp3")?;
-///     assert!(infer::is_audio_read(&mut f).unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_audio_read<R>(r: &mut R) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_audio_read(r)
-}
-
 /// Determines whether a buffer is a book type.
 ///
 /// # Examples
@@ -885,29 +523,6 @@ where
 /// ```
 pub fn is_book(buf: &[u8]) -> bool {
     INFER.is_book(buf)
-}
-
-/// Determines whether data from buffer is a book type.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.epub")?;
-///     assert!(infer::is_book_read(&mut f).unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_book_read<R>(r: &mut R) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_book_read(r)
 }
 
 /// Determines whether a buffer is a document type.
@@ -922,29 +537,6 @@ pub fn is_document(buf: &[u8]) -> bool {
     INFER.is_document(buf)
 }
 
-/// Determines whether data from reader is a document type.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.docx")?;
-///     assert!(infer::is_document_read(&mut f).unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_document_read<R>(r: &mut R) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_document_read(r)
-}
-
 /// Determines whether a buffer is a font type.
 ///
 /// # Examples
@@ -955,29 +547,6 @@ where
 /// ```
 pub fn is_font(buf: &[u8]) -> bool {
     INFER.is_font(buf)
-}
-
-/// Determines whether data from reader is a font type.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.ttf")?;
-///     assert!(infer::is_font_read(&mut f).unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_font_read<R>(r: &mut R) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_font_read(r)
 }
 
 /// Determines whether a buffer is an image type.
@@ -992,29 +561,6 @@ pub fn is_image(buf: &[u8]) -> bool {
     INFER.is_image(buf)
 }
 
-/// Determines whether data from reader is an image type.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.png")?;
-///     assert!(infer::is_image_read(&mut f).unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_image_read<R>(r: &mut R) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_image_read(r)
-}
-
 /// Determines whether a buffer is a video type.
 ///
 /// # Examples
@@ -1025,29 +571,6 @@ where
 /// ```
 pub fn is_video(buf: &[u8]) -> bool {
     INFER.is_video(buf)
-}
-
-/// Determines whether data from reader is a video type.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::fs;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let mut f = File::open("testdata/sample.mov")?;
-///     assert!(infer::is_video_read(&mut f).unwrap());
-///     Ok(())
-/// }
-/// ```
-#[cfg(feature = "std")]
-pub fn is_video_read<R>(r: &mut R) -> io::Result<bool>
-where
-    R: Read + Seek,
-{
-    INFER.is_video_read(r)
 }
 
 /// Returns the file type for the mime type if supported.
